backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.4
sqlmodel==0.0.22
psycopg[binary]==3.2.1
asyncpg==0.29.0
python-dotenv==1.0.1
authlib==1.3.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
python-multipart==0.0.9
requests==2.32.3
aiofiles==24.1.0
httpx==0.27.2
redis==5.0.8
celery==5.4.0
langchain==0.3.7
langchain-community==0.3.7
ollama==0.1.9
pypdf==5.1.0
python-docx==1.1.2
tenacity==9.0.0
structlog==24.2.0
